Extract loading and error comps in common ones
